backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.11
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.33.0
python-dotenv==1.0.0
rapidfuzz==2.15.1
psycopg2-binary==2.9.6
